Added error checking so that you don't pop and empty stack, and fixed some formating of outputs for easier readability.

+++ stack.cpp
@@ -4,7 +4,6 @@
  */
 
 #include <iostream>
-#include <assert.h>
 #include "stack.h"
 using namespace std;
 
@@ -23,8 +22,6 @@ Stack::~Stack()
    }
 }
 
-//\end{verbatim}  \lecpb  \begin{verbatim}
-
 void Stack::Push(int n)
 {
    topPtr.AddNode( n );
@@ -33,14 +30,15 @@ void Stack::Push(int n)
 
 int Stack::Pop()
 {
-   assert(!IsEmpty());
-
-   int  n = topPtr.FirstNode();
-   topPtr.DeleteNode( n );
-   return n;
+    if(!IsEmpty()){
+        int  n = topPtr.FirstNode();
+        topPtr.DeleteNode( n );
+        return n;
+    } else {
+        return 0;
+    }
 }
 
-//\end{verbatim}  \lecpb  \begin{verbatim}
 
 int Stack::IsEmpty()
 {
